requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
sqlalchemy>=2.0
pdfplumber>=0.11
openpyxl>=3.1
reportlab>=4.1
python-dotenv>=1.0
